requirements.txt
# Core ML dependencies
torch>=2.1.0
transformers>=4.38.0
datasets>=2.15.0
accelerate>=0.27.0
bitsandbytes>=0.42.0

# Scientific computing
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0

# Configuration and utilities
PyYAML>=6.0
pydantic>=2.0.0
tqdm>=4.60.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.28.0

# System monitoring
psutil>=5.8.0
nvidia-ml-py3>=7.352.0

# HuggingFace ecosystem
huggingface-hub>=0.16.0

# Development dependencies (optional)
# pytest
# pre-commit
# black
# isort
# flake8
# mypy
